+++ i18n/en.pot
@@ -5,8 +5,14 @@ msgstr ""
 "Content-Type: text/plain; charset=utf-8\n"
 "Content-Transfer-Encoding: 8bit\n"
 "Plural-Forms: nplurals=2; plural=(n != 1)\n"
-"POT-Creation-Date: 2022-03-25T15:58:50.753Z\n"
-"PO-Revision-Date: 2022-03-25T15:58:50.753Z\n"
+"POT-Creation-Date: 2022-03-25T23:21:44.950Z\n"
+"PO-Revision-Date: 2022-03-25T23:21:44.950Z\n"
+
+msgid "Name is already taken"
+msgstr "Name is already taken"
+
+msgid "Cancel"
+msgstr "Cancel"
 
 msgid "User Guide"
 msgstr "User Guide"
